claude: refactor(kaspa): move relayer module from libs/kaspa to rust/main
- Move relayer source files to rust/main/chains/dymension-kaspa/src/kas_relayer/
- Update imports to use dym_kas_core, dym_kas_bridge, dym_kas_hardcode
- Remove relayer from libs/kaspa workspace
- Update hyperlane-base to use dymension_kaspa::kas_relayer

This reduces the HL dependencies in libs/kaspa, working towards the goal
of keeping only pure Kaspa code there.
