+++ dymension/libs/kaspa/tooling/Cargo.toml
@@ -9,7 +9,6 @@ corelib = { path = "../lib/core", package = "core" }
 bridge = { path = "../lib/bridge" }
 hardcode = { path = "../lib/hardcode" }
 validator = { path = "../lib/validator" }
-relayer = { path = "../lib/relayer" }
 secp256k1 = { workspace = true }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = { workspace = true }
