+++ rust/main/chains/dymension-kaspa/src/kas_relayer/mod.rs
@@ -9,4 +9,4 @@ pub use withdraw::messages::on_new_withdrawals;
 // Re-export metrics for easier access
 pub use metrics::KaspaBridgeMetrics;
 
-pub use secp256k1::PublicKey;
+pub use kaspa_bip32::secp256k1::PublicKey;
