+++ rust/main/chains/dymension-kaspa/src/mailbox.rs
@@ -3,7 +3,7 @@ use crate::withdrawal_utils::{
     calculate_failed_indexes, record_withdrawal_batch_metrics, WithdrawalStage,
 };
 use crate::KaspaProvider;
-use dym_kas_relayer::withdraw::minimum::is_small_value;
+use crate::kas_relayer::withdraw::minimum::is_small_value;
 use hyperlane_core::{
     utils::bytes_to_hex, BatchResult, ChainResult, ContractLocator, Decode, FixedPointNumber,
     HyperlaneChain, HyperlaneContract, HyperlaneDomain, HyperlaneMessage, HyperlaneProvider,
